Fix undefined error

+++ database/migrations/2021_11_12_211829_create_mediafiles_table.php
@@ -17,8 +17,8 @@ class CreateMediafilesTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('type');
-            $table->integer('size');
-            $table->integer('duration');
+            $table->bigInteger('size');
+            $table->bigInteger('duration');
             $table->integer('fps');
             $table->timestamp('modified_at');
             $table->timestamps();
